Feature exposure (Pearson) in evaluator
