+++ tests/test_evaluation.py
@@ -6,7 +6,7 @@ from unittest.mock import patch
 from numerblox.misc import Key
 from numerblox.evaluation import NumeraiClassicEvaluator, NumeraiSignalsEvaluator
 
-from utils import create_signals_sample_data
+from utils import create_signals_sample_data, classic_test_data
 
 BASE_STATS_COLS = ['target', 'mean', 'std', 'sharpe', 'max_drawdown', 'apy', 'calmar_ratio', 'corr_with_example_preds', 'legacy_mean', 'legacy_std', 'legacy_sharpe', 'max_feature_exposure', 'feature_neutral_mean', 'feature_neutral_std', 'feature_neutral_sharpe', 'tb200_mean', 'tb200_std', 'tb200_sharpe', 'tb500_mean', 'tb500_std', 'tb500_sharpe', 'exposure_dissimilarity']
 CLASSIC_SPECIFIC_STATS_COLS = ['feature_neutral_mean_v3', 'feature_neutral_std_v3', 'feature_neutral_sharpe_v3']
@@ -14,10 +14,8 @@ CLASSIC_SPECIFIC_STATS_COLS = ['feature_neutral_mean_v3', 'feature_neutral_std_v
 CLASSIC_STATS_COLS = BASE_STATS_COLS + CLASSIC_SPECIFIC_STATS_COLS
 SIGNALS_STATS_COLS = BASE_STATS_COLS
 
-df = create_signals_sample_data
-
-def test_numerai_classic_evaluator():
-    df = pd.read_parquet("tests/test_assets/train_int8_5_eras.parquet")
+def test_numerai_classic_evaluator(classic_test_data):
+    df = classic_test_data
     df.loc[:, "prediction"] = np.random.uniform(size=len(df))
     df.loc[:, "prediction_random"] = np.random.uniform(size=len(df))
 
@@ -31,7 +29,8 @@ def test_numerai_classic_evaluator():
     for col in CLASSIC_STATS_COLS + CLASSIC_SPECIFIC_STATS_COLS:
         assert col in val_stats.columns
 
-def test_numerai_signals_evaluator(df):
+def test_numerai_signals_evaluator(create_signals_sample_data):
+    df = create_signals_sample_data
     evaluator = NumeraiSignalsEvaluator(era_col="date", fast_mode=False)
     val_stats = evaluator.full_evaluation(
         dataf=df,
@@ -57,7 +56,8 @@ def mock_api():
         mock_instance.diagnostics.return_value = [{"status": "done", "perEraDiagnostics": [{"era": "2023-09-01", "validationCorr": 0.6}]}]
         yield mock_instance
 
-def test_get_neutralized_corr(df, mock_api):
+def test_get_neutralized_corr(create_signals_sample_data, mock_api):
+    df = create_signals_sample_data
     obj = NumeraiSignalsEvaluator(era_col="date", fast_mode=True)  
     result = obj.get_neutralized_corr(df, "test_model", Key("Hello", "World"))
     
@@ -77,8 +77,8 @@ def test_await_diagnostics_timeout(mock_api):
     with pytest.raises(Exception, match=r"Diagnostics couldn't be retrieved within .* minutes after uploading."):
         obj._NumeraiSignalsEvaluator__await_diagnostics(api=mock_api, model_id="test_model_id", diagnostics_id="test_diag_id", timeout_min=0.001, interval_sec=2)
 
-def test_numerai_classic_meta_model_corr():
-    df = pd.read_parquet("tests/test_assets/train_int8_5_eras.parquet")
+def test_numerai_classic_meta_model_corr(classic_test_data):
+    df = classic_test_data
     df.loc[:, "prediction"] = np.random.uniform(size=len(df))
     df.loc[:, "prediction_random"] = np.random.uniform(size=len(df))
 
@@ -95,3 +95,18 @@ def test_numerai_classic_meta_model_corr():
 
     assert val_stats["corr_with_meta_model"].min() >= -1
     assert val_stats["corr_with_meta_model"].max() <= 1
+
+def test_get_raw_feature_exposures(classic_test_data):
+    evaluator = NumeraiClassicEvaluator(era_col="era", fast_mode=False)
+    np.random.seed(1)
+    classic_test_data["prediction"] = np.random.uniform(size=len(classic_test_data))
+
+    feature_list = [col for col in classic_test_data.columns if col.startswith("feature")]
+    raw_exposures = evaluator.get_raw_feature_exposures(classic_test_data, pred_col="prediction", feature_list=feature_list)
+    assert isinstance(raw_exposures, pd.DataFrame)
+    assert len(raw_exposures) == len(classic_test_data["era"].unique())
+    # Check that values are between -1 and 1
+    assert raw_exposures.min().min() >= -1
+    assert raw_exposures.max().max() <= 1
+    for feature in feature_list:
+        assert feature in raw_exposures.columns
